This is an odd change
Randomize used to be able to null out enums, it no longer does this as
it's considered a 'formatted' value which do not do nulls.

They also used to be i%3 (general) and i%2 (enum) so the odds of
becoming null were even different. This led to a low probability of a
null string and an enum both being null, or winding up with the same
values etc.

The problem that happened then is now that there's no null enums every once
in a while there would be insufficient random changes made to the
event_ones/twos model in the mysql test schema for Update/Upsert calls.
This caused mysql to return RowsAffected = 0 and the test would fail.

In order to descrease the chance of getting the exact same model
randomized twice, I've combined event_ones, event_twos, and event_threes
because I have no idea what they're trying to accomplish by being
together.