Add context parameter throughout sqlboiler.
This was accompanied by a syntactically breaking change to the way we do
queries. We no longer store a handle to the database on or around any
query. Instead queries at the time of execution will have to be passed
the database handle. This is actually much more consistent despite the
original approach having been for consistency.

- Add interfaces for context based execution.
- Add flag --no-context to disable context parameter generation/use.
- Add context.Context to all templates and function calls.
- Move the boil.Executor parameter to the query execution site.
- Remove the storage of boil.Executor inside queries.