Remove additional reliance on the null package
We now have the ability to use custom types everywhere. The null package
could be replaced or removed or what have you. But there were subtle hints
everywhere of it's existence. These assumptions have been removed and
in their place we assume that types were are dealing with will implement
sql.Scanner/driver.Valuer so that we can assign and read their values.

This comes with some small performance penalties because more
reflection. Shouldn't be noticeable though.

- Add some functions to help assign/compare and generally work with the
  sql.Scanner and driver.Valuer interfaces.
- Remove all references to null package in the templates.