Make NonZeroDefaultSet look at struct tags
Now that things can be aliased, the struct tag checking is more
important for anything looking for struct fields using the actual column
name (which this method does).

We also stopped doing TitleCase when constructing field maps, since
everything is always lowercased anyway, and we have struct tags on all
structs we care about, let the poor performance and edge-case-yness of
the unTitleCase method rest on the people who do not choose to use the
struct tag as the fix is simple (use the struct tag).

- Struct mappings no longer TitleCase everywhere
- NonZeroDefaultSet is now aware of struct tags because field names
  don't always correspond to their respective column name anymore.