Fix a bug in mysql upsert
A long long time ago, someone opened PR #177 trying to resolve an issue
with mysql upsert. I made a fix, but I never actually pulled it into any
mainline code for some reason (commit 23723353). This is the same change
made to the v3 code.

The reason for this change is because mysql can conflict automatically
during upsert on any unique column. This makes total sense. However if
it does happen to conflict on an insert for which we don't have an ID
(which is probably all the time), then when we go to retrieve the record
it fails because we don't have an id and we try to look it up using said
id. Instead now we look for non-zero unique columns as the way to find
the record (which includes the id if it's non-zero).

- Fix mysql upsert bug reported in PR #177