Bindata templates
This change is being implemented not only because it was requested but
because it was the source of a number of bugs. The old behavior of
basedir was sneaky as it would find a GOPATH on your system you didn't
mean for it to find and dig out old templates and cause problems.

- Add better debugging output for "where this file comes from"
- Add bindata template loading
- Fix #151