Add alternative syntax for aliases
When creating this syntax, I thought about being "safe" like I was with
the syntax for replacements, because toml keys are fairly restrictive.
But I decided that no one would have odd characters in their table
names. And no one has, yet. But apparently viper downcases all keys that
it reads from toml, even though the toml library in use doesn't appear
to do this.

The good thing is it resolves to the exact same data structure in the
end anyway so no other changes were necessary.

- Add test for new syntax
- Document new syntax
- Contribute towards fixing #180