Unify Bind() result and error handling with v3
A bug was recently fixed with Bind() where it would consume the entire
result set when asked to bind to a single object. However the for loop
always completed by .Next() returning false. This implicitly closed the
rows. Now that we break at the correct time for .One() type queries this
bug was introduced.

We deferred the close of the rows, so that result set is still opened
when eager loading begins. This caused problems with psql transactions
when eager loading.

Also, awkwardly enough this wasn't a bug in v3 because it was handling
things differently, so we simply make the code the same and it should
fix it nicely.

- Explicitly close rows in queries.Bind()
- Fix #318