enable support for IDENTITY
PostgreSQL 10.2+ introduces IDENTITY columns as drop-in replacement for SERIAL. This fix allow identifying identities and mark them as with defaults.

Before doing this, we check if PostgreSQL information_schema has is_identity field. If not then false is returned. This provides backward-compatibility for older versions of PostgreSQL that does not have this field.