Add conversion in randomize
If the type is an alias of say an integer, the int case will fire and
create a random int, but when we try to assign that it fails. So we do a
check to see if the types are equal to each other, if not we attempt the
conversion before assignment.

- Fix #487