Fix for the extremely slow model generation with some MySQL databases.
If one tries to generate the models on a MySQL database with a lot of
databases and tables the generation is getting very slow (#329).

This seems to result from MySQL not optimizing queries on views very well.

The solution I have chosen and validated is to replace the table and schema
in the subquery with the actual names instead of references to the parent
queries.

The result for my system is insane and went down from over 9 hours to 4
seconds.

This fix is kinda ugly but straight forward and I think the speed advantage
is worth it. Especially as the code is never being seen by anyone.