Change nullability checks for where helper types
This should fix a situation where a type is first seen without being
null, then later it's seen as being null but does not generate null
helpers.

The downside is this will add null helpers to things that are both null
and not nullable. But this is maybe a separate concern.