Add getter methods to relationship structs
The goal of this change is to alleviate the need for the caller to check
if the models' relationship struct is nil before accessing a field.
Indeed, we've seen a proliferation in our codebase of the following
pattern:

```
if record.R != nil && record.R.SomeField != nil {
   ...
}
```

This is especially required in utilities functions which are not
responsible for loading the model and can't know if the record has
eagerly loaded the relationships and thus can't know if the relationship
struct is nil. With this change, the previous code idiom can be replaced
with:

```
if field := record.R.GetSomeField(); field != nil {
  ...
}

```

The important part is that there's no possibility for panicing due to
nil pointer, where-as, previously,  if the caller forgot to check for
`record.R != nil`, a panic would happen.

The change is backward compatible and follows the pattern that protobuf
generation uses, i.e. `Message.GetX() XType`.