Fix qm.WithDeleted helper with a custom soft delete column
As the `deletedAtRgx` was hardcoded, there wasn't way to use
the `qm.WithDeleted` helper when you have a custom soft delete
column set in `auto-columns.deleted` config.

This PR fixes that, overwriting the variable when the the
custom column is set and the soft deletion capatibilies are
activated for that table.

Closes #1256