Downgrade decimal version before decompose interface
Any version after https://github.com/ericlagergren/decimal/pull/124
causes problems when saving Decimal to sql drivers like postgres. While
the best solution may be to update all the sql drivers, it has been 3
years since then, so it is better to make sure the latest version does
not have that commit. This version is the latest before it.